Bugfixes for Read/Write of profiles and records.  All should be working now.

+++ SeventyTwoDesktop/Controllers/FileReadWriteController.cs
@@ -2,14 +2,23 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Threading;
+
 using System.Threading.Tasks;
 using System.IO;
 using System.Timers;
 
 namespace SeventyTwoDesktop.Controllers
 {
+
+
     class FileReadWriteController
     {
+
+        public static readonly object WriteLocker = new object( );
+
+        public static readonly object TimerLocker = new object( );
+
         public FileReadWriteController( string path, string backupPath = "", double interval = 5000 ) {
             FileName = path;
             TimerInterval = interval;
@@ -24,7 +33,7 @@ namespace SeventyTwoDesktop.Controllers
         private string FileName { get; set; } = string.Empty;
         private string BackupFileName { get; set; } = string.Empty;
         private string FileContentsToWrite { get; set; } = string.Empty;
-        private Timer MainTimer { get; set; }
+        private System.Timers.Timer MainTimer { get; set; }
         private double TimerInterval { get; set; }
         private bool Dirty { get; set; } = false;
         private bool WriteSuccess { get; set; }
@@ -32,11 +41,14 @@ namespace SeventyTwoDesktop.Controllers
         public bool LastWriteSucceeded { get { return WriteSuccess; } }
         public string FileContents { get { return FileContentsToWrite; } }
         public string TargetFile { get { return FileName; } }
-        
+
+
+        private Guid gu { get; set; } = Guid.NewGuid( );
+
         private void InitTimer() {
             //If we don't have a timer, start one.
             if( MainTimer == null ) {
-                MainTimer = new Timer( TimerInterval );
+                MainTimer = new System.Timers.Timer( TimerInterval );
                 MainTimer.Elapsed += new ElapsedEventHandler( TimerEvent ); 
                 MainTimer.Start( );
                 IntervalsSinceLastWrite = 0;
@@ -50,7 +62,9 @@ namespace SeventyTwoDesktop.Controllers
         private void TimerEvent( object ElapsedEventArgs, EventArgs e ) {
             //If the data has changed, we definitely need to write it.
             if ( Dirty ) {
-                PerformWrite( );
+                lock( TimerLocker ){
+                    PerformWrite( );
+                }
             //If we have had 5 intervals since we last wrote, we should stop this timer and wait for another text event.
             } else if ( IntervalsSinceLastWrite >= 5 ) {
                 MainTimer.Stop( );
@@ -58,26 +72,35 @@ namespace SeventyTwoDesktop.Controllers
             } else {
                 IntervalsSinceLastWrite++;
             }
+            
         }
 
         private void PerformWrite() {
 
             try {
-                //Should we be backing this up?
-                if( !string.IsNullOrEmpty( BackupFileName ) ) {
-                    //Delete an existing duplicate.
-                    if( File.Exists( BackupFileName ) ) {
-                        File.Delete( BackupFileName );
-                    }
-                    //Make a quick copy of the old list
-                    if( File.Exists( FileName ) ) {
-                        File.Copy( FileName, BackupFileName );
+
+                if( Dirty ) {
+                    lock( WriteLocker ) {
+                        //Should we be backing this up?
+                        if( !string.IsNullOrEmpty( BackupFileName ) ) {
+                            //Delete an existing duplicate.
+                            if( File.Exists( BackupFileName ) ) {
+                                File.Delete( BackupFileName );
+                            }
+                            //Make a quick copy of the old list
+                            if( File.Exists( FileName ) ) {
+                                File.Copy( FileName, BackupFileName );
+                            }
+                        }
+
+                        File.WriteAllText( FileName, FileContentsToWrite );
+
+                        Dirty = false;
+                        WriteSuccess = true;
+                        IntervalsSinceLastWrite = 0;
                     }
                 }
-                File.WriteAllText( FileName, FileContentsToWrite );
-                Dirty = false;
-                WriteSuccess = true;
-                IntervalsSinceLastWrite = 0;
+
             } catch ( Exception exc ) { Models.Log.WriteToLog( exc ); }
         }
